requirements.txt
chainlit==1.1.402
langchain==0.3.1
langchain-groq==0.2.0
langchain-community==0.3.1
faiss-cpu==1.9.0
pypdf2==3.0.1
python-dotenv==1.0.1
sentence-transformers==3.2.0